continue refactorings
